+++ lib/flutter_share.dart
@@ -15,18 +15,18 @@ class FlutterShare {
   /// - LinkUrl: Is the [linkUrl] to include with the message.
   /// - ChooserTitle (Just for Android): Is the [chooserTitle] of the app
   /// chooser popup. If null the system default title will be used.
-  static Future<bool> share(
-      {@required String title,
-      String text,
-      String linkUrl,
-      String chooserTitle}) async {
-    assert(title != null && title.isNotEmpty);
+  static Future<bool?> share(
+      {required String title,
+      String? text,
+      String? linkUrl,
+      String? chooserTitle}) async {
+    assert(title.isNotEmpty);
 
-    if (title == null || title.isEmpty) {
+    if (title.isEmpty) {
       throw FlutterError('Title cannot be null');
     }
 
-    final bool success = await _channel.invokeMethod('share', <String, dynamic>{
+    final success = await _channel.invokeMethod('share', <String, dynamic>{
       'title': title,
       'text': text,
       'linkUrl': linkUrl,
@@ -38,28 +38,30 @@ class FlutterShare {
 
   /// Shares a local file with the app chooser.
   ///
-  /// - Title: Is the [title] of the message. Used as email subject if sharing
+  /// - Title: It's the [title] of the message. Used as email subject if sharing
   /// with mail apps. The [title] cannot be null.
-  /// - Text: Is the [text] of the message.
-  /// - FilePath: Is the [filePath] to include with the message.
-  /// - ChooserTitle (Just for Android): Is the [chooserTitle] of the app
-  /// chooser popup. If null the system default title will be used.
-  static Future<bool> shareFile(
-      {@required String title,
-      @required String filePath,
-      String text,
-      String chooserTitle,
+  /// - Text: It's the [text] of the message.
+  /// - FilePath: It's the [filePath] to include with the message.
+  /// - ChooserTitle (Just for Android): It's the [chooserTitle] of the app
+  /// chooser popup. If null, the system default title will be used.
+  /// - FileType (Just for Android): It's the [fileType] that will be sent in the 
+  /// chooser popup. If null, the system default title will be used.
+  static Future<bool?> shareFile(
+      {required String title,
+      required String filePath,
+      String? text,
+      String? chooserTitle,
       String fileType = '*/*'}) async {
-    assert(title != null && title.isNotEmpty);
-    assert(filePath != null && filePath.isNotEmpty);
+    assert(title.isNotEmpty);
+    assert(filePath.isNotEmpty);
 
-    if (title == null || title.isEmpty) {
+    if (title.isEmpty) {
       throw FlutterError('Title cannot be null');
-    } else if (filePath == null || filePath.isEmpty) {
+    } else if (filePath.isEmpty) {
       throw FlutterError('FilePath cannot be null');
     }
 
-    final bool success =
+    final success =
         await _channel.invokeMethod('shareFile', <String, dynamic>{
       'title': title,
       'text': text,